chore: split into fetch/ and scan/ monorepo directories
Each container now has its own Dockerfile, pyproject.toml, requirements,
pipeline package, and tests. CI splits into path-filtered fetch.yml and
scan.yml workflows — a change to fetch/ only triggers the fetch pipeline.
