--- .github/workflows/ci.yml
@@ -1,151 +1,0 @@
-name: CI
-
-on:
-  push:
-    branches: [main]
-  pull_request:
-    branches: [main]
-
-jobs:
-  test:
-    name: Unit Tests
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v6
-
-      - name: Set up Python
-        uses: actions/setup-python@v6
-        with:
-          python-version: "3.12"
-
-      - name: Install dev dependencies
-        run: pip install -r requirements-dev.txt
-
-      - name: Run tests
-        run: pytest
-
-  build-and-check:
-    name: Build, Health Check & Push
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-      packages: write
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v6
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v4
-
-      - name: Log in to GitHub Container Registry
-        if: github.event_name == 'push' && github.ref == 'refs/heads/main'
-        uses: docker/login-action@v4
-        with:
-          registry: ghcr.io
-          username: ${{ github.actor }}
-          password: ${{ secrets.GITHUB_TOKEN }}
-
-      # --- fetch image ---
-
-      - name: Extract metadata (fetch)
-        id: meta-fetch
-        uses: docker/metadata-action@v6
-        with:
-          images: ghcr.io/${{ github.repository_owner }}/music-pipeline-fetch
-          tags: |
-            type=raw,value=latest,enable=${{ github.ref == 'refs/heads/main' }}
-            type=sha,format=short
-
-      - name: Build fetch image
-        uses: docker/build-push-action@v7
-        with:
-          context: .
-          target: fetch
-          push: false
-          load: true
-          tags: |
-            ${{ steps.meta-fetch.outputs.tags }}
-            music-pipeline-fetch:ci-local
-          labels: ${{ steps.meta-fetch.outputs.labels }}
-          cache-from: type=gha
-          cache-to: type=gha,mode=max
-
-      - name: Health check — spotdl
-        run: docker run --rm --entrypoint spotdl music-pipeline-fetch:ci-local --version
-
-      - name: Health check — fetch commands
-        run: |
-          docker run --rm music-pipeline-fetch:ci-local sh -c "
-            which music-ingest &&
-            which music-setup &&
-            which music-provision
-          "
-
-      - name: Push fetch image to GHCR
-        if: github.event_name == 'push' && github.ref == 'refs/heads/main'
-        uses: docker/build-push-action@v7
-        with:
-          context: .
-          target: fetch
-          push: true
-          tags: ${{ steps.meta-fetch.outputs.tags }}
-          labels: ${{ steps.meta-fetch.outputs.labels }}
-          cache-from: type=gha
-          cache-to: type=gha,mode=max
-
-      # --- scan image ---
-
-      - name: Extract metadata (scan)
-        id: meta-scan
-        uses: docker/metadata-action@v6
-        with:
-          images: ghcr.io/${{ github.repository_owner }}/music-pipeline-scan
-          tags: |
-            type=raw,value=latest,enable=${{ github.ref == 'refs/heads/main' }}
-            type=sha,format=short
-
-      - name: Build scan image
-        uses: docker/build-push-action@v7
-        with:
-          context: .
-          target: scan
-          push: false
-          load: true
-          tags: |
-            ${{ steps.meta-scan.outputs.tags }}
-            music-pipeline-scan:ci-local
-          labels: ${{ steps.meta-scan.outputs.labels }}
-          cache-from: type=gha
-          cache-to: type=gha,mode=max
-
-      - name: Health check — beet
-        run: docker run --rm --entrypoint beet music-pipeline-scan:ci-local version
-
-      - name: Health check — ffmpeg
-        run: docker run --rm --entrypoint ffmpeg music-pipeline-scan:ci-local -version
-
-      - name: Health check — fpcalc
-        run: docker run --rm --entrypoint fpcalc music-pipeline-scan:ci-local -version
-
-      - name: Health check — scan commands
-        run: |
-          docker run --rm music-pipeline-scan:ci-local sh -c "
-            which music-scan &&
-            which music-import &&
-            which music-remove
-          "
-
-      - name: Push scan image to GHCR
-        if: github.event_name == 'push' && github.ref == 'refs/heads/main'
-        uses: docker/build-push-action@v7
-        with:
-          context: .
-          target: scan
-          push: true
-          tags: ${{ steps.meta-scan.outputs.tags }}
-          labels: ${{ steps.meta-scan.outputs.labels }}
-          cache-from: type=gha
-          cache-to: type=gha,mode=max
