+++ .github/workflows/fetch.yml
@@ -1,0 +1,94 @@
+name: fetch CI
+
+on:
+  push:
+    branches: [main]
+    paths: [fetch/**, compose.yml, justfile]
+  pull_request:
+    branches: [main]
+    paths: [fetch/**, compose.yml, justfile]
+
+jobs:
+  test:
+    name: Unit Tests (fetch)
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v6
+
+      - name: Set up Python
+        uses: actions/setup-python@v6
+        with:
+          python-version: "3.12"
+
+      - name: Install dependencies
+        working-directory: fetch
+        run: pip install -r requirements.txt -r requirements-dev.txt -e .
+
+      - name: Run tests
+        working-directory: fetch
+        run: pytest
+
+  build-and-check:
+    name: Build, Health Check & Push (fetch)
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+      packages: write
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v6
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v4
+
+      - name: Log in to GitHub Container Registry
+        if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+        uses: docker/login-action@v4
+        with:
+          registry: ghcr.io
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Extract metadata (fetch)
+        id: meta
+        uses: docker/metadata-action@v6
+        with:
+          images: ghcr.io/${{ github.repository_owner }}/music-pipeline-fetch
+          tags: |
+            type=raw,value=latest,enable=${{ github.ref == 'refs/heads/main' }}
+            type=sha,format=short
+
+      - name: Build fetch image
+        uses: docker/build-push-action@v7
+        with:
+          context: fetch
+          target: prod
+          push: false
+          load: true
+          tags: |
+            ${{ steps.meta.outputs.tags }}
+            music-pipeline-fetch:ci-local
+          labels: ${{ steps.meta.outputs.labels }}
+          cache-from: type=gha
+          cache-to: type=gha,mode=max
+
+      - name: Health check — spotdl
+        run: docker run --rm --entrypoint spotdl music-pipeline-fetch:ci-local --version
+
+      - name: Health check — fetch entry points
+        run: docker run --rm music-pipeline-fetch:ci-local sh -c "which music-ingest"
+
+      - name: Push fetch image to GHCR
+        if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+        uses: docker/build-push-action@v7
+        with:
+          context: fetch
+          target: prod
+          push: true
+          tags: ${{ steps.meta.outputs.tags }}
+          labels: ${{ steps.meta.outputs.labels }}
+          cache-from: type=gha
+          cache-to: type=gha,mode=max
