+++ .github/workflows/scan.yml
@@ -1,0 +1,100 @@
+name: scan CI
+
+on:
+  push:
+    branches: [main]
+    paths: [scan/**, compose.yml, justfile]
+  pull_request:
+    branches: [main]
+    paths: [scan/**, compose.yml, justfile]
+
+jobs:
+  test:
+    name: Unit Tests (scan)
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v6
+
+      - name: Set up Python
+        uses: actions/setup-python@v6
+        with:
+          python-version: "3.12"
+
+      - name: Install dependencies
+        working-directory: scan
+        run: pip install -r requirements.txt -r requirements-dev.txt -e .
+
+      - name: Run tests
+        working-directory: scan
+        run: pytest
+
+  build-and-check:
+    name: Build, Health Check & Push (scan)
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+      packages: write
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v6
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v4
+
+      - name: Log in to GitHub Container Registry
+        if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+        uses: docker/login-action@v4
+        with:
+          registry: ghcr.io
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Extract metadata (scan)
+        id: meta
+        uses: docker/metadata-action@v6
+        with:
+          images: ghcr.io/${{ github.repository_owner }}/music-pipeline-scan
+          tags: |
+            type=raw,value=latest,enable=${{ github.ref == 'refs/heads/main' }}
+            type=sha,format=short
+
+      - name: Build scan image
+        uses: docker/build-push-action@v7
+        with:
+          context: scan
+          target: prod
+          push: false
+          load: true
+          tags: |
+            ${{ steps.meta.outputs.tags }}
+            music-pipeline-scan:ci-local
+          labels: ${{ steps.meta.outputs.labels }}
+          cache-from: type=gha
+          cache-to: type=gha,mode=max
+
+      - name: Health check — beet
+        run: docker run --rm --entrypoint beet music-pipeline-scan:ci-local version
+
+      - name: Health check — ffmpeg
+        run: docker run --rm --entrypoint ffmpeg music-pipeline-scan:ci-local -version
+
+      - name: Health check — fpcalc
+        run: docker run --rm --entrypoint fpcalc music-pipeline-scan:ci-local -version
+
+      - name: Health check — scan entry points
+        run: docker run --rm music-pipeline-scan:ci-local sh -c "which music-scan && which music-import"
+
+      - name: Push scan image to GHCR
+        if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+        uses: docker/build-push-action@v7
+        with:
+          context: scan
+          target: prod
+          push: true
+          tags: ${{ steps.meta.outputs.tags }}
+          labels: ${{ steps.meta.outputs.labels }}
+          cache-from: type=gha
+          cache-to: type=gha,mode=max
