--- requirements-dev.txt
@@ -1,3 +1,0 @@
--r requirements-fetch.txt
--r requirements-scan.txt
-pytest>=8.0
