+++ scan/pipeline/music_pipeline.py
