+++ scan/pyproject.toml
@@ -3,14 +3,13 @@ requires = ["setuptools>=68"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "music-pipeline"
+name = "music-pipeline-scan"
 version = "0.1.0"
 requires-python = ">=3.11"
 dependencies = []
 
 [project.scripts]
 music-scan   = "pipeline.cli:scan"
-music-ingest = "pipeline.cli:ingest"
 music-import = "pipeline.cli:import_cmd"
 
 [tool.setuptools.packages.find]
