+++ scan/tests/test_music_pipeline.py
